+++ crates/wolf_engine_codegen/src/lib.rs
@@ -2,11 +2,11 @@ use proc_macro::{self, TokenStream};
 use quote::quote;
 use syn::{parse_macro_input, DeriveInput};
 
-#[proc_macro_derive(Event)]
-pub fn derive_event(input: TokenStream) -> TokenStream {
+#[proc_macro_derive(DynamicEvent)]
+pub fn derive_dynamic_event(input: TokenStream) -> TokenStream {
     let DeriveInput { ident, .. } = parse_macro_input!(input);
     let output = quote! {
-        impl Event for #ident {}
+        impl DynamicEvent for #ident {}
     };
     output.into()
 }