+++ crates/wolf_engine_events/src/dynamic/event.rs
@@ -1,18 +1,17 @@
-//! Provides dynamically-typed events for the engine.
+pub use wolf_engine_codegen::DynamicEvent;
 
 use downcast_rs::*;
+
 use std::fmt::Debug;
 
-pub use wolf_engine_codegen::Event;
-
-/// Represents a [`Boxed`](Box) dynamic [`Event`].
-pub type EventBox = Box<dyn Event>;
+/// Represents a [`Boxed`](Box) [`DynamicEvent`].
+pub type DynamicEventBox = Box<dyn DynamicEvent>;
 
 /// A dynamically-typed event.
 ///
 /// Events can be downcasted back to their original type using the [`Downcast`] trait.
-pub trait Event: Downcast + Debug + 'static {}
-impl_downcast!(Event);
+pub trait DynamicEvent: Downcast + Debug + 'static {}
+impl_downcast!(DynamicEvent);
 
 #[cfg(test)]
 mod event_tests {
@@ -20,11 +19,11 @@ mod event_tests {
 
     use super::*;
 
-    #[derive(Event, Debug)]
+    #[derive(DynamicEvent, Debug)]
     struct TestEvent(&'static str);
 
     #[test_case(&TestEvent("Hello, World!"))]
-    fn should_auto_impl_event(event: &dyn Event) {
+    fn should_auto_impl_event(event: &dyn DynamicEvent) {
         if let Some(event) = event.downcast_ref::<TestEvent>() {
             assert_eq!(event.0, "Hello, World!");
         }