+++ crates/wolf_engine_events/src/dynamic/mod.rs
@@ -1,0 +1,9 @@
+//! Provides dynamically-typed events for the engine.
+
+mod event;
+pub use event::*;
+
+pub mod event_loop;
+
+mod event_sender;
+pub use event_sender::*;